Add custom User model
A custom user model has been created per [Django's docs](https://docs.djangoproject.com/en/2.1/topics/auth/customizing/#substituting-a-custom-user-model). The model is part of the accounts app, which has also been created.

NOTE: This is a breaking change to the database. Since it is still early in development, it's easiest to drop the DB / tables.